Add moving platforms - which of course turned into a spaghetti nightmare
There are quite a few other things in here as well like a refactor of how enemies damage the player and various interfaces to attempt to clean things up a bit while making them more flexible for the future. I realized I needed to solve a lot of things before trying to handle all the edge cases of moving platforms. Not sure everything is solved still, but it all seems fairly decent at least. There is quite some spaghetti here though. Hopefully it can be cleaned up more in the future.

+++ Assets/Spelunky/Scripts/Entity/EntityVisuals.cs
@@ -5,7 +5,7 @@ namespace Spelunky {
     public class EntityVisuals : MonoBehaviour {
 
         public bool isFacingRight = false;
-        public float facingDirection = -1;
+        public int facingDirection = -1;
 
         // References.
         [HideInInspector] public new SpriteRenderer renderer;
